+++ app/Services/Users/UserAvatarHandlerService.php
@@ -10,7 +10,8 @@ use App\Models\User;
 class UserAvatarHandlerService
 {
     public function __construct(
-        protected UserAvatarService $avatarService
+        protected UserAvatarService $avatarService,
+        protected AvatarValidator $validator
     ) {
     }
 
@@ -24,11 +25,10 @@ class UserAvatarHandlerService
      */
     public function handleUpload(User $user, mixed $avatar): void
     {
-        if (! $this->isValidUpload($avatar)) {
-            return;
+        if ($this->validator->isValidUpload($avatar)) {
+            $path = $this->avatarService->upload($avatar, $user);
+            $user->update(['avatar' => $path]);
         }
-
-        $this->uploadAndSave($user, $avatar);
     }
 
     /**
@@ -41,13 +41,14 @@ class UserAvatarHandlerService
      */
     public function handleUpdate(User $user, mixed $avatar): void
     {
-        if ($this->shouldRemove($avatar)) {
+        if ($this->validator->shouldRemove($avatar)) {
             $this->remove($user);
             return;
         }
 
-        if ($this->isValidUpload($avatar)) {
-            $this->replaceAndSave($user, $avatar);
+        if ($this->validator->isValidUpload($avatar)) {
+            $path = $this->avatarService->replace($avatar, $user);
+            $user->update(['avatar' => $path]);
         }
     }
 
@@ -62,57 +63,5 @@ class UserAvatarHandlerService
     {
         $this->avatarService->delete($user);
         $user->update(['avatar' => null]);
-    }
-
-    /**
-     * Check if avatar should be removed.
-     *
-     * @param  mixed $avatar
-     *
-     * @return bool
-     */
-    private function shouldRemove(mixed $avatar): bool
-    {
-        return $avatar === null || $avatar === '';
-    }
-
-    /**
-     * Check if avatar is a valid upload.
-     *
-     * @param  mixed $avatar
-     *
-     * @return bool
-     */
-    private function isValidUpload(mixed $avatar): bool
-    {
-        return is_object($avatar) && method_exists($avatar, 'isValid');
-    }
-
-    /**
-     * Upload avatar and save to user.
-     *
-     * @param  User $user
-     * @param  mixed $avatar
-     *
-     * @return void
-     */
-    private function uploadAndSave(User $user, mixed $avatar): void
-    {
-        $path = $this->avatarService->upload($avatar, $user);
-        $user->update(['avatar' => $path]);
-    }
-
-    /**
-     * Replace avatar and save to user.
-     *
-     * @param  User $user
-     * @param  mixed $avatar
-     *
-     * @return void
-     */
-    private function replaceAndSave(User $user, mixed $avatar): void
-    {
-        $path = $this->avatarService->replace($avatar, $user);
-        $user->update(['avatar' => $path]);
     }
 }
